refactor(oidc): clean up in the train
